unify SageMakerEndpoint and SageMakerEndpointWithStream into SageMaker LLM
