--- docs/extras/modules/model_io/models/llms/integrations/aws_sagemaker_with_stream.mdx
@@ -1,14 +1,0 @@
-# AWS SageMakerEndpointWithStream
-
-LangChain.js supports integration with AWS SageMaker-hosted endpoints also using streams. Check [Amazon SageMaker JumpStart](https://aws.amazon.com/sagemaker/jumpstart/) for a list of available models, and how to deploy your own.
-
-Here's an example:
-
-```bash npm2yarn
-npm install @aws-sdk/client-sagemaker-runtime
-```
-
-import CodeBlock from "@theme/CodeBlock";
-import SageMakerEndpointWithStreamExample from "@examples/models/llm/sagemaker_endpoint_with_stream.ts";
-
-<CodeBlock language="typescript">{SageMakerEndpointWithStreamExample}</CodeBlock>
