+++ examples/.env.example
@@ -56,3 +56,6 @@ ANALYTICDB_USERNAME=ADD_YOURS_HERE
 ANALYTICDB_PASSWORD=ADD_YOURS_HERE
 ANALYTICDB_DATABASE=ADD_YOURS_HERE
 ASSEMBLYAI_API_KEY=ADD_YOURS_HERE
+AWS_ACCESS_KEY_ID=ADD_YOURS_HERE
+AWS_SECRET_ACCESS_KEY=ADD_YOURS_HERE
+AWS_REGION=ADD_YOURS_HERE
