+++ langchain/.env.example
@@ -71,3 +71,4 @@ ASSEMBLYAI_API_KEY=ADD_YOURS_HERE
 LLAMA_PATH=ADD_YOURS_HERE
 AWS_ACCESS_KEY_ID=ADD_YOURS_HERE
 AWS_SECRET_ACCESS_KEY=ADD_YOURS_HERE
+AWS_REGION=ADD_YOURS_HERE
