+++ langchain/.gitignore
@@ -136,12 +136,9 @@ llms/googlepalm.d.ts
 llms/fireworks.cjs
 llms/fireworks.js
 llms/fireworks.d.ts
-llms/sagemaker_endpoint.cjs
-llms/sagemaker_endpoint.js
-llms/sagemaker_endpoint.d.ts
-llms/sagemaker_endpoint_with_stream.cjs
-llms/sagemaker_endpoint_with_stream.js
-llms/sagemaker_endpoint_with_stream.d.ts
+llms/sagemaker.cjs
+llms/sagemaker.js
+llms/sagemaker.d.ts
 llms/bedrock.cjs
 llms/bedrock.js
 llms/bedrock.d.ts
