+++ langchain/scripts/create-entrypoints.js
@@ -60,8 +60,7 @@ const entrypoints = {
   "llms/googlevertexai/web": "llms/googlevertexai/web",
   "llms/googlepalm": "llms/googlepalm",
   "llms/fireworks": "llms/fireworks",
-  "llms/sagemaker_endpoint": "llms/sagemaker_endpoint",
-  "llms/sagemaker_endpoint_with_stream": "llms/sagemaker_endpoint_with_stream",
+  "llms/sagemaker": "llms/sagemaker",
   "llms/bedrock": "llms/bedrock",
   "llms/llama_cpp": "llms/llama_cpp",
   "llms/writer": "llms/writer",
@@ -292,8 +291,7 @@ const requiresOptionalDependency = [
   "llms/hf",
   "llms/raycast",
   "llms/replicate",
-  "llms/sagemaker_endpoint",
-  "llms/sagemaker_endpoint_with_stream",
+  "llms/sagemaker",
   "llms/bedrock",
   "llms/llama_cpp",
   "llms/writer",
