+++ langchain/src/load/import_constants.ts
@@ -27,8 +27,7 @@ export const optionalImportEntrypoints = [
   "langchain/llms/googlevertexai",
   "langchain/llms/googlevertexai/web",
   "langchain/llms/googlepalm",
-  "langchain/llms/sagemaker_endpoint",
-  "langchain/llms/sagemaker_endpoint_with_stream",
+  "langchain/llms/sagemaker",
   "langchain/llms/bedrock",
   "langchain/llms/llama_cpp",
   "langchain/llms/writer",
