+++ langchain/src/load/import_type.d.ts
@@ -79,12 +79,9 @@ export interface OptionalImportMap {
   "langchain/llms/googlepalm"?:
     | typeof import("../llms/googlepalm.js")
     | Promise<typeof import("../llms/googlepalm.js")>;
-  "langchain/llms/sagemaker_endpoint"?:
-    | typeof import("../llms/sagemaker_endpoint.js")
-    | Promise<typeof import("../llms/sagemaker_endpoint.js")>;
-  "langchain/llms/sagemaker_endpoint_with_stream"?:
-    | typeof import("../llms/sagemaker_endpoint_with_stream.js")
-    | Promise<typeof import("../llms/sagemaker_endpoint_with_stream.js")>;
+  "langchain/llms/sagemaker"?:
+    | typeof import("../llms/sagemaker.js")
+    | Promise<typeof import("../llms/sagemaker.js")>;
   "langchain/llms/bedrock"?:
     | typeof import("../llms/bedrock.js")
     | Promise<typeof import("../llms/bedrock.js")>;
