+++ langchain/tsconfig.json
@@ -78,8 +78,7 @@
       "src/llms/googlevertexai/web.ts",
       "src/llms/googlepalm.ts",
       "src/llms/fireworks.ts",
-      "src/llms/sagemaker_endpoint.ts",
-      "src/llms/sagemaker_endpoint_with_stream.ts",
+      "src/llms/sagemaker.ts",
       "src/llms/bedrock.ts",
       "src/llms/llama_cpp.ts",
       "src/llms/writer.ts",
